+++ core/src/elements/ons-toolbar.js
@@ -48,6 +48,7 @@ const scheme = {
  *   [/en]
  *   [ja]ナビゲーションで使用するツールバー用コンポーネントです。クラス名により、左、中央、右のコンテナを指定できます。[/ja]
  * @codepen aHmGL
+ * @tutorial vanilla/Reference/button
  * @guide Addingatoolbar [en]Adding a toolbar[/en][ja]ツールバーの追加[/ja]
  * @seealso ons-bottom-toolbar
  *   [en]The `<ons-bottom-toolbar>` displays a toolbar on the bottom of the page.[/en]