+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 CHANGELOG
 ====
 
+v2.0.0-rc.6
+----
+ * ons-splitter: 'load' methods return promises in Angular1 bindings.
+ * ons-navigator: Fix 'popPage' with 'refresh' parameter.
+ * ons-dialog: Make ons-dialog angular2 compatible.
+ * ons-speed-dial, ons-speed-dial-item: Removed `_compiled` attribute.
+ * ons-speed-dial: Make ons-speed-dial elements angular2 compatible.
+
 v2.0.0-rc.5
 ----
- * ons-dialog: Make ons-dialog angular2 compatible.
- * ons-speed-dial, ons-speed-dial-item: Removed "_compiled" attribute.
- * ons-speed-dial: Make ons-speed-dial elements angular2 compatible.
  * css-components: Fix material list item paddings.
  * ons-list: Fix [#1401](https://github.com/OnsenUI/OnsenUI/issues/1401).
  * ons-range: Fix [#1391](https://github.com/OnsenUI/OnsenUI/issues/1391).
@@ -15,6 +20,11 @@ v2.0.0-rc.5
  * ons-speed-dial: Emit `open` and `close` events.
  * ons-speed-dial: Block clicks when hidden.
  * ons-splitter: Correctly import dependency.
+ * css-components: Fix border of inline list.
+ * css-components: Refine style of list item title and subtitle.
+ * ons-toolbar: Fix [#1419](https://github.com/OnsenUI/OnsenUI/issues/1419).
+ * css-components: Refine default paddings of list items.
+ * ons-back-button: Add `onClick` property.
 
 v2.0.0-rc.4
 ----